- Getting server to be inline with postman collection

+++ api/migrations/0001_initial.py
@@ -21,7 +21,10 @@ class Migration(migrations.Migration):
                     verbose_name='ID')
                  ),
                 ('name', models.CharField(max_length=256)),
-                ('price', models.FloatField()),
+                ('price', models.DecimalField(
+                    decimal_places=2,
+                    max_digits=16
+                )),
             ],
         ),
     ]
